// src/app/shared/models/planet.ts
export class Planet {
    name: string;
    rotation_period: number;
    orbital_period: number;
    diameter: number;
    climate: string;
    gravity: string;
    terrain: string;
    surface_water: number;
    population: string;
    residents: any[];
    films: any[];
    created: Date;
    edited:Date;
    url: string;
}